feat: add empty document ui, fix patch api
